libs/aion-server-langgraph/pyproject.toml
[tool.poetry]
name = "aion-server-langgraph"
version = "0.1.0"
description = "A2A server wrapper for LangGraph projects"
authors = ["Terminal Research Team <support@terminal.exchange>"]
readme = "README.md"
packages = [{include = "aion", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12.0"
aion-api-client = { git = "https://github.com/Terminal-Research/aion-python-sdk", branch = "main", subdirectory = "libs/aion-api-client" }
uvicorn = "0.23.2"
fastapi = "0.119.1"
structlog = "24.4.0"
httpx = "0.28.1"
# PostgreSQL
psycopg = { version = "3.2.9", extras = ["binary"] }
asyncpg = "0.30.0" # native psql protocol (faster then psycopg)
greenlet = "3.2.3" # supports async pg operations
alembic = "1.16.2" # migrations
SQLAlchemy = "2.0.41"

langgraph-checkpoint-postgres = "2.0.21"
langchain-openai = "0.3.27"
PyYAML = "6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
